App factory refactor (#10)
* Move registering blueprint to separate function

* Move all extensions init to extensions.py

* Move extensions init to a separate function

* Fix import statements after moving extension init to extensions.py
