+++ app/db/__init__.py
@@ -1,5 +1,1 @@
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
 
-db = SQLAlchemy()
-migrate = Migrate(directory='db/migrations')
