+++ app/extensions.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+from flask_rq2 import RQ
+
+db = SQLAlchemy()
+migrate = Migrate(directory='db/migrations')
+rq = RQ()
