+++ app/jobs/__init__.py
@@ -1,8 +1,3 @@
-from flask_rq2 import RQ
-
-rq = RQ()
-
-
 def update_job_status(job, status):
     job.meta['status'] = status
     job.save_meta()
